Make morale affect speed in imps.

+++ view_object.h
@@ -55,7 +55,7 @@ class ViewObject {
   ViewObject& setAttribute(Attribute, double);
   optional<float> getAttribute(Attribute) const;
 
-  using CreatureAttributes = EnumMap<AttrType, std::uint8_t>;
+  using CreatureAttributes = EnumMap<AttrType, std::uint16_t>;
   void setCreatureAttributes(CreatureAttributes);
   const optional<CreatureAttributes>& getCreatureAttributes() const;
